tidied up filter display code

+++ src/components/selectedFiltersDisplay.js
@@ -46,61 +46,44 @@ const FormatBtnLabel = ({category, filter}) => {
 const SelectedFiltersDisplay = () => {
     const { readOnlyFilters, filterCount } = useSelectedFiltersState();
     const { deleteFilter, deleteAllFilters } = useSelectedFiltersUpdate();
+    
+    if (filterCount < 1) {
+      return <div className="font-light italic">Add some filters below</div>
+    }
 
     return (
       <>
         <h2 className="text-2xl font-bold tracking-wide mb-5">Current Filters</h2>
 
-        {
-          filterCount > 0
-          ?   
+        {  
             <>
               <div className="flex space-between flex-wrap my-4">
-
-               { /* TODO neaten this up */
-                  Object.keys(readOnlyFilters).map(key => {
-                    if (Array.isArray(readOnlyFilters[key])) {
-
-                      return readOnlyFilters[key].map( (filter, i) => (
-                        <span key={`filter-${key}_${i}`} className="mr-2 mb-2">
+               { 
+                  Object.entries(readOnlyFilters).map( ([category, filterList] ) => (
+                      filterList.map( (filter, i) => (
+                        <span key={`filter-${category}_${i}`} className="mr-2 mb-2">
                         
                         <Button 
                           variant="pill" 
-                          onClick={ () => deleteFilter(key, filter)}
+                          onClick={ () => deleteFilter(category, filter)}
                           icon={<>&#215;</>}
                         >
-                          <FormatBtnLabel category={key} filter={filter} />
+                          <FormatBtnLabel category={category} filter={filter} />
                         </Button>
 
                         </span>
                       ))
-
-                    } else {
-                      return (
-                        <span className="mr-2 mb-2">
-                        
-                        <Button 
-                          variant="pill" 
-                          onClick={ () => deleteFilter(key, readOnlyFilters[key])}
-                          icon={<>&#215;</>}
-                        >
-                          <FormatBtnLabel category={key} filter={readOnlyFilters[key]} />
-                        </Button>
-
-                        </span>
-                      )}
-                  })
+                  ))
                 }
               </div>
+              
               <div className="flex w-64 justify-between">
                 <p className="text-semibold"><strong>{filterCount}</strong>{` filter${filterCount > 1 ? 's' :'' } selected`}</p>
                 <Button variant="link" onClick={() => deleteAllFilters()}> 
                   Clear all filters
                 </Button>
               </div>
-        
             </>
-          : <div className="font-light italic">Add some filters below</div>
         }
         
       </>
